Update
More novice tiles, this brings the total to 60. Made some script changes
to AvatarController.cs to allow for movement to be adjusted for large
tail and avatar sizes
Avatar is no 10% and tails are now 20% larger.

Fixed an issue where to score was incrementing at on in the first tile,
now the score stays at 0 until you pass the first tile.