Update
Modified the speed of the avatar and tested, landing on 10% slower as a
good speed.

Changed the variable of the avatar speed to a float from an int.

Challenge room tiles now only appear every 7 tiles.

Fixed issue with Challenge room tiles where the doors spawned on top of
the boarders.

General Tile adjustments.